ajustando xmlwos_run pois existem alguns pids inválidos, no momento apenas validando tamanho da string do PID

+++ isis2mongodb/sh/xmlwos_run.sh
@@ -54,15 +54,19 @@ do
     loaded=`curl -s -X GET "http://"$scielo_data_url"/api/v1/is_loaded?code="$pid`
     if [[ $loaded == "False" ]]; then
         mkdir -p ../output/isos/$pid
-        $cisis_dir/mx ../../databases/isis/artigo  btell="0" pid=$pid   iso=../output/isos/$pid/$pid"_artigo.iso" -all now
-        $cisis_dir/mx ../../databases/isis/title   btell="0" ${pid:1:9} iso=../output/isos/$pid/$pid"_title.iso" -all now
-        $cisis_dir/mx ../../databases/isis/bib4cit btell="0" $pid"$"    iso=../output/isos/$pid/$pid"_bib4cit.iso" -all now
+        issn=${pid:1:9}
+        len=${#pid}
+        if [[ $len -eq 23 ]]; then
+            $cisis_dir/mx ../../databases/isis/artigo  btell="0" pid=$pid   iso=../output/isos/$pid/$pid"_artigo.iso" -all now
+            $cisis_dir/mx ../../databases/isis/title   btell="0" $issn      iso=../output/isos/$pid/$pid"_title.iso" -all now
+            $cisis_dir/mx ../../databases/isis/bib4cit btell="0" $pid"$"    iso=../output/isos/$pid/$pid"_bib4cit.iso" -all now
 
-        python isis2json.py ../output/isos/$pid/$pid"_artigo.iso" -c -p v -t 3 > ../output/isos/$pid/$pid"_artigo.json"
-        python isis2json.py ../output/isos/$pid/$pid"_title.iso" -c -p v -t 3 > ../output/isos/$pid/$pid"_title.json"
-        python isis2json.py ../output/isos/$pid/$pid"_bib4cit.iso" -c -p v -t 3 > ../output/isos/$pid/$pid"_bib4cit.json"
-        curl -X POST "http://"$scielo_data_url"/api/v1/article?code="$pid
-        rm -rf ../output/isos/$pid/*.iso    
+            python isis2json.py ../output/isos/$pid/$pid"_artigo.iso" -c -p v -t 3 > ../output/isos/$pid/$pid"_artigo.json"
+            python isis2json.py ../output/isos/$pid/$pid"_title.iso" -c -p v -t 3 > ../output/isos/$pid/$pid"_title.json"
+            python isis2json.py ../output/isos/$pid/$pid"_bib4cit.iso" -c -p v -t 3 > ../output/isos/$pid/$pid"_bib4cit.json"
+            curl -X POST "http://"$scielo_data_url"/api/v1/article?code="$pid
+            rm -rf ../output/isos/$pid/*.iso
+        fi
     else
         echo "article alread processed!!!"
     fi
